added a model-folder and a NetflixTitles.js so I can import it as a component. Also added a rule so that if fields like title, director, cast and country are empty, they will be filled with unknown
